fix renaming all to singular requests name and fixed tests
