+++ tests/test_bill_requests.py
@@ -3,8 +3,8 @@ import sys, os
 path = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0,path+'/../')
 
-from nessie.accountRequests import AccountRequests
-from nessie import billRequests
+from nessie.accountRequest import AccountRequest
+from nessie.billRequest import BillRequest
 from nessie.dataRequest import DataRequest
 from nessie.utils.exceptions import NessieApiError
 
@@ -23,7 +23,7 @@ account_id = '5a5471796514d52c7774a2cb'
 class TestBillRequests(unittest.TestCase):
     # create some dummy bills
     def setUp(self):
-        bill_factory = billRequests.BillRequest(wkey)
+        bill_factory = BillRequest(wkey)
         bill = bill_factory.create_bill(account_id,
             status='pending',
             payee='bobby',
@@ -40,7 +40,7 @@ class TestBillRequests(unittest.TestCase):
     # try fetching a bill that doesn't exist
     def test_get_nonreal_bill_fail(self):
         print("test_get_nonreal_bill_fail")
-        bill_factory = billRequests.BillRequest(wkey)
+        bill_factory = BillRequest(wkey)
         
         
         expected_status_code = 404
